refactoring(2단계 - 로또(자동)) : 도메인 수정
- 로또 결과도 도메인에서 도출 하도록 수정
