+++ src/main/java/lotto/domain/LottoTickets.java
@@ -2,27 +2,27 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
-import static lotto.domain.LottoNumber.getAutoNumberList;
+import static lotto.domain.LottoNumbers.getAutoNumberList;
 
 public class LottoTickets {
 
     private final PurchasePrice purchasePrice;
     private List<Lotto> lottoList;
 
-    public LottoTickets(PurchasePrice purchasePrice, List<Lotto> lottoList) {
-        this(purchasePrice);
-        this.lottoList = lottoList;
-    }
-
     public LottoTickets(PurchasePrice purchasePrice) {
         this.purchasePrice = purchasePrice;
+    }
+
+    public List<Lotto> getLottoList() {
+        return this.lottoList;
     }
 
     public LottoTickets pickNumbers() {
         List<Lotto> lottoList = new ArrayList<>();
 
-        for (int i = 0; i < purchasePrice.getAmount(); i++) {
+        for (int i = 0; i < this.purchasePrice.getAmount(); i++) {
             lottoList.add(new Lotto(getAutoNumberList()));
         }
 
@@ -30,17 +30,24 @@ public class LottoTickets {
         return this;
     }
 
-    public Integer getPurchasePrice() {
-        return purchasePrice.getPurchasePrice();
-    }
-
     public LottoTickets putRankings(List<Integer> matchNumberList) {
-        lottoList.forEach(l -> l.rank(matchNumberList));
+        this.lottoList.forEach(l -> l.rank(matchNumberList));
         return this;
     }
 
-    public List<Lotto> getLottoList() {
-        return lottoList;
+    public List<Rank> getRank() {
+        return this.lottoList
+                .stream()
+                .map(Lotto::getRank)
+                .collect(Collectors.toList());
+    }
+
+    public Integer getYield() {
+        int totalWinningMoney = this.lottoList
+                .stream()
+                .mapToInt(l -> l.getRank().getWinningMoney())
+                .sum();
+        return totalWinningMoney / this.purchasePrice.getPurchasePrice();
     }
 
 }
